challenge 1: the variable sum is initialized at "sum=0"
the looping condition is “array < arrayOf
Numbers.length” which states that the loop will run until the value of “array” is less than the length of arrayOfNumber. After each iteration, the value of “array” would be incremented by “1“.

the statement inside the for loop updates the value of the sum variable after each iteration by adding the array element in it

then return the sum to the function after everything passes.

challenge 2: